Add padding & redefine types

+++ tests/utils.c
@@ -6,35 +6,35 @@
 /*   By: ohakola <ohakola@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/27 17:57:02 by ohakola           #+#    #+#             */
-/*   Updated: 2021/02/27 17:58:45 by ohakola          ###   ########.fr       */
+/*   Updated: 2021/02/27 18:47:37 by ohakola          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "utils.h"
 
-void			rand_array(int32_t *array, int32_t array_size)
+void			rand_array(uint32_t *array, uint32_t array_size)
 {
-	int32_t	i;
+	size_t	i;
 
 	i = -1;
 	while (++i < array_size)
 		array[i] = rand();
 }
 
-void			copy_array(int32_t *dst, int32_t *src,
-					int32_t arrays_size)
+void			copy_array(uint32_t *dst, uint32_t *src,
+					size_t arrays_size)
 {
-	int32_t	i;
+	size_t	i;
 
 	i = -1;
 	while (++i < arrays_size)
 		dst[i] = src[i];
 }
 
-t_bool			arrays_match(int32_t *arr1, int32_t *arr2,
-					int32_t arrays_size)
+t_bool			arrays_match(uint32_t *arr1, uint32_t *arr2,
+					size_t arrays_size)
 {
-	int32_t	i;
+	size_t	i;
 
 	i = -1;
 	while (++i < arrays_size)
@@ -45,11 +45,11 @@ t_bool			arrays_match(int32_t *arr1, int32_t *arr2,
 
 int				check_comp(const void *a, const void *b) 
 {
-	int32_t	f;
-	int32_t	s;
+	size_t	f;
+	size_t	s;
 
-	f = *((int*)a);
-	s = *((int*)b);
+	f = *((int32_t*)a);
+	s = *((int32_t*)b);
 	if (f > s)
 		return  (1);
 	if (f < s)
@@ -57,7 +57,7 @@ int				check_comp(const void *a, const void *b)
 	return (0);
 }
 
-void			sort_check_array(int32_t *check_array, int32_t array_size)
+void			sort_check_array(uint32_t *check_array, uint32_t array_size)
 {
 	qsort(check_array, array_size, sizeof(*check_array), check_comp);
 }
